+++ angular/src/bsb-theme/css/style.css
@@ -10578,3 +10578,12 @@ html.ie11 .bs-searchbox .form-control {
 thead{
   background: #fff;
 }
+
+.custom-scrollbar::-webkit-scrollbar {
+  width: 7px;
+  background-color: #F5F5F5;
+}
+.custom-scrollbar::-webkit-scrollbar-thumb {
+  background-color: #cecece;
+  border: 2px solid #cecece;
+}
